# docs/swagger.yaml
basePath: /api/v1
definitions:
  controllers.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  controllers.SuccessResponse:
    properties:
      message:
        example: operation successful
        type: string
    type: object
  controllers.User:
    description: 用戶模型
    properties:
      email:
        example: zhangsan@example.com
        type: string
      id:
        example: 507f1f77bcf86cd799439011
        type: string
      name:
        example: 張三
        type: string
    required:
    - email
    - name
    type: object
host: localhost:8080
info:
  contact: {}
  description: 這是一個使用 Gin 和 MongoDB 的 RESTful API 服務
  title: Go API with Gin and MongoDB
  version: "1.0"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: 獲取系統中的所有用戶列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 獲取所有用戶
      tags:
      - users
    post:
      consumes:
      - application/json
      description: 創建一個新的用戶
      parameters:
      - description: 用戶信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controllers.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 創建新用戶
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: 刪除特定用戶
      parameters:
      - description: 用戶ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 刪除用戶
      tags:
      - users
    get:
      consumes:
      - application/json
      description: 通過ID獲取特定用戶的信息
      parameters:
      - description: 用戶ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 獲取特定用戶
      tags:
      - users
    put:
      consumes:
      - application/json
      description: 更新特定用戶的信息
      parameters:
      - description: 用戶ID
        in: path
        name: id
        required: true
        type: string
      - description: 用戶信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controllers.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: 更新用戶
      tags:
      - users
produces:
- application/json
schemes:
- http
swagger: "2.0"